Swap BrowserRouter for HashRouter
The reason why we're doing this change is so that we support direct
links.  the trick is the hash in the Url: as React apps get handled
client-side and SSR is a whole other history, the HashRouter is best
suited for it. The reason why this works is that the /#/ tells the
server "ignore everything from the hash onwards; everything's gonna be
handled client side".