Adds APY to wrap component
This adds the APY metric that was added to useUnipoolRewards to the
Token carousel. In order to do so it was necessary to refactor how
Tokens were configured.

Previously a global object modeAttributes was used to configure carousel
items by passing a key as a propoerty. However, the APY metric depends
on useUnipoolRewards so it couldn't be added to the global object.

Now the Token mode property expects an object rather than a key, and the
object is passed from the WrapToken component to the Token componet.

Previously the key was also being used to conditionaly render "Earned"
when using the "claim" mode, since we are not passing a key, this now
depends on the the label value of the dictoary.